requirements.txt
streamlit
pandas
numpy
xgboost
joblib
requests
plotly
optuna
scikit-learn
pydeck
python-dotenv
